[DSD-231] Rename existing checkbox to allow for development of new one.
