+++ src/components/Checkbox-old/CheckboxOld.stories.tsx
@@ -1,12 +1,12 @@
 import React, { useEffect, useState } from "react";
-import Checkbox, { CheckboxProps } from "./Checkbox";
+import CheckboxOld, { CheckboxProps } from "./CheckboxOld";
 import { Story } from "@storybook/react/types-6-0";
 import { withDesign } from "storybook-addon-designs";
 
 // Set defaults for the checkbox stories.
 export default {
-  title: "Checkbox",
-  component: Checkbox,
+  title: "Checkbox-old",
+  component: CheckboxOld,
   // Remove some props for all checkbox stories.
   argTypes: {
     checkboxId: { table: { disable: true } },
@@ -17,7 +17,7 @@ export default {
   decorators: [withDesign],
 };
 
-const Template: Story<CheckboxProps> = args => <Checkbox {...args} />;
+const Template: Story<CheckboxProps> = args => <CheckboxOld {...args} />;
 
 // Uncontrolled checkbox.
 export const UncontrolledCheckbox = Template.bind({});
@@ -51,7 +51,7 @@ const ControlledCheckboxTemplate: Story<CheckboxProps> = args => {
   }, [args.checked]);
 
   return (
-    <Checkbox
+    <CheckboxOld
       {...args}
       checkboxId={"checkbox"}
       checked={isChecked}
