+++ src/styles.scss
@@ -26,7 +26,7 @@
 @import "./components/Accordion/_Accordion.scss";
 @import "./components/Breadcrumbs/_Breadcrumbs.scss";
 @import "./components/Card/_Card.scss";
-@import "./components/Checkbox/_Checkbox.scss";
+@import "./components/Checkbox-old/_CheckboxOld.scss";
 @import "./components/HelperErrorText/_HelperErrorText.scss";
 @import "./components/Hero/_Hero.scss";
 @import "./components/Image/_Image.scss";
